Added C++ and Python functions for creating motions and skeletons procedurally (i.e. without having to load a motion or skeleton file). To create a skeleton in Python (C++ is similar): sk = assets.createSkeletonDefinition("myskeleton") j = sk.createJoint("myjoint", None) j2 = sk.createJoint("anotherjoint", j)
To create a motion in Python (C++ is similar):
m = assets.createMotion("mymotion")
m.addChannel("base", "XPos")
m.addChannel("base", "YPos")
data = FloatVec()
data.append(1.4)
data.append(3.2)
m.addFrame(0, data)

cleaned up some warnings

+++ core/smartbody/SmartBody/src/sb/SBJointMap.cpp
@@ -809,7 +809,7 @@ bool SBJointMap::guessMapping(SmartBody::SBSkeleton* skeleton, bool prtMap)
 	if(spine4 && spine4->num_children()>0)
 	{
 		std::vector<SkJoint*> j_list;
-		for (unsigned int i=0;i<spine4->num_children();i++)
+		for (unsigned int i=0;i< (size_t) spine4->num_children();i++)
 		{
 			SkJoint* j = spine4->child(i);
 			SrString jname(j->jointName().c_str());
